modify car

+++ app/clientView.py
@@ -30,6 +30,10 @@ def clientView():
             
         if 'return' in request.form:
             return redirect(url_for('user.root'))
+        
+        if 'modify' in request.form:
+            session['client_id'] = request.form['modify']
+            return redirect(url_for('clientView.modifyClient'))
 
         if 'find' in request.form:
             find = request.form['find']
@@ -95,6 +99,71 @@ def addClient():
         flash(error)
     return render_template('index/analist/addClient.html')
 
+@bp.route('/modifyClient', methods=('POST', 'GET'))
+def modifyClient():
+    db = get_db()
+    id = session['client_id']
+
+    if request.method == 'POST':
+        dni = request.form['dni']
+        firstname = request.form['firstname']
+        secondname = request.form['secondname']
+        phone = request.form['phone']
+        mail = request.form['mail']
+        address = request.form['address']
+
+        db = get_db()
+        error = None
+
+        if not re.search("^[\w\.]+@([\w-]+\.)+[\w-]{2,4}$", mail):
+            error = 'You must enter a valid mail'
+
+        if not re.search("^[V|E]-\d+$", dni):
+            error = 'Invalid DNI'
+
+        if not re.search("^\d{4}-\d{7}$", phone):
+            error = 'Invalid phone number'
+
+        if not re.search("^[\w\.]+@([\w-]+\.)+[\w-]{2,4}$", mail):
+            error = 'You must enter a valid mail'
+
+        if not re.search("^[V|E]-\d+$", dni):
+            error = 'Invalid DNI'
+
+        if not re.search("^\d{4}-\d{7}$", phone):
+            error = 'Invalid phone number'
+
+        if error is None:
+            try:
+                db.execute("""UPDATE clients SET dni = ? WHERE id = ?""", 
+                            (dni, id))
+                
+                db.execute("""UPDATE clients SET firstname = ? WHERE id = ?""", 
+                            (firstname, id))
+                
+                db.execute("""UPDATE clients SET secondname = ? WHERE id = ?""", 
+                            (secondname, id))
+
+                db.execute("""UPDATE clients SET tlf = ? WHERE id = ?""", 
+                            (phone, id))
+                
+                db.execute("""UPDATE clients SET mail = ? WHERE id = ?""", 
+                            (mail, id))
+                
+                db.execute("""UPDATE clients SET address = ? WHERE id = ?""", 
+                            (address, id))
+                
+                db.commit()
+
+                return redirect(url_for('clientView.clientView'))
+            
+            except db.IntegrityError:
+                error = 'Client already registered'
+        flash(error)
+    client = db.execute("""SELECT * FROM clients WHERE id = ?""", (id)).fetchone()
+
+    return render_template('index/analist/modifyClient.html', client=client)
+
 @bp.route('/details', methods=('POST', 'GET'))
 def clientDetail():
     db = get_db()
@@ -107,13 +176,17 @@ def clientDetail():
         
         if 'return' in request.form:
             return redirect(url_for('clientView.clientView'))
+        
+        if 'modify' in request.form:
+            session['car_id'] = request.form['modify']
+            return redirect(url_for('clientView.modifyCar'))
 
         if 'delete' in request.form:
             id = request.form['delete']
             db.execute("""
                         DELETE FROM cars WHERE id = ?""", id)
             db.commit()
-
+        
         if 'find' in request.form:
             find = request.form['find']
 
@@ -182,3 +255,67 @@ def addCar():
 
         flash(error)
     return render_template('index/analist/addCar.html')
+
+@bp.route('/modifyCar', methods=('POST', 'GET'))
+def modifyCar():
+    id = session['car_id']
+    db = get_db()
+
+    if request.method == "POST":
+        plaque = request.form['plaque']
+        brand = request.form['brand']
+        model = request.form['model']
+        year = request.form['year']
+        bodywork = request.form['bodywork']
+        motor = request.form['motor']
+        color = request.form['color']
+        problem = request.form['problem']
+
+        db = get_db()
+        error = None
+
+        if not re.search("[\d\w]{17}", bodywork):
+            error = 'Invalid bodywork serial'
+
+        elif not re.search("[\d\w]{17}", motor):
+            error = 'Invalid motor serial'
+
+        elif not re.search("\d{4}", year) or int(year) > 2023:
+            error = 'Invalid year'
+
+        if error is None:
+            try:
+                db.execute("""UPDATE cars SET plaque = ? WHERE id = ?""", 
+                            (plaque, id))
+                
+                db.execute("""UPDATE cars SET brand = ? WHERE id = ?""", 
+                            (brand, id))
+                
+                db.execute("""UPDATE cars SET model = ? WHERE id = ?""", 
+                            (model, id))
+                
+                db.execute("""UPDATE cars SET year = ? WHERE id = ?""", 
+                            (year, id))
+                
+                db.execute("""UPDATE cars SET bodyWorkSerial = ? WHERE id = ?""", 
+                            (bodywork, id))
+                
+                db.execute("""UPDATE cars SET motorSerial = ? WHERE id = ?""", 
+                            (motor, id))
+                
+                db.execute("""UPDATE cars SET color = ? WHERE id = ?""", 
+                            (color, id))
+                
+                db.execute("""UPDATE cars SET problem = ? WHERE id = ?""", 
+                            (problem, id))
+                db.commit()
+
+                return redirect(url_for('clientView.clientDetail'))
+            
+            except db.IntegrityError:
+                error = 'Car already registered'
+
+        flash(error)
+
+    car = db.execute("SELECT * FROM cars WHERE id = ?", id).fetchone()
+    return render_template('index/analist/modifyCar.html', car=car)
